Add timeout option to fetchers
This uses each fetchers specific timeout ENV vars and command
arguments, and it is up to those programs themselves to honor
that timeout. For instance: there is currently no guarantee
that setting the timeout for `git` to 1 will actually timeout
after 1 second.

Ideally we'd be able to set timeouts on Procs directly, but once
Proc::Async works on JVM again we can switch back to it which will
make things much easier.